requirements.txt
asgiref==3.5.2
Django~=3.2.9
psycopg2-binary==2.9.5
sqlparse==0.4.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
djangorestframework~=3.14.0